# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Sketch {
  id: number;
  title: string;
  imageDataUrl: string;
  generatedCode: string;
  framework: string;
  /** @nullable */
  instructions?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type SketchInputFramework = typeof SketchInputFramework[keyof typeof SketchInputFramework];


export const SketchInputFramework = {
  'react-tailwind': 'react-tailwind',
  'react-shadcn': 'react-shadcn',
  'html-tailwind': 'html-tailwind',
} as const;

export interface SketchInput {
  title: string;
  imageDataUrl: string;
  framework: SketchInputFramework;
  instructions?: string;
}

export type RegenerateInputFramework = typeof RegenerateInputFramework[keyof typeof RegenerateInputFramework];


export const RegenerateInputFramework = {
  'react-tailwind': 'react-tailwind',
  'react-shadcn': 'react-shadcn',
  'html-tailwind': 'html-tailwind',
} as const;

export interface RegenerateInput {
  instructions?: string;
  framework?: RegenerateInputFramework;
}

export type SketchStatsByFramework = {[key: string]: number};

export interface SketchStats {
  total: number;
  byFramework: SketchStatsByFramework;
  recentCount: number;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}