# lib/api-zod/src/generated/types/regenerateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegenerateInputFramework } from './regenerateInputFramework';

export interface RegenerateInput {
  instructions?: string;
  framework?: RegenerateInputFramework;
}