# lib/api-zod/src/generated/types/sketch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Sketch {
  id: number;
  title: string;
  imageDataUrl: string;
  generatedCode: string;
  framework: string;
  /** @nullable */
  instructions?: string | null;
  createdAt: Date;
  updatedAt: Date;
}