# lib/api-zod/src/generated/types/sketchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SketchInputFramework } from './sketchInputFramework';

export interface SketchInput {
  title: string;
  imageDataUrl: string;
  framework: SketchInputFramework;
  instructions?: string;
}