# lib/api-zod/src/generated/types/sketchStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SketchStatsByFramework } from './sketchStatsByFramework';

export interface SketchStats {
  total: number;
  byFramework: SketchStatsByFramework;
  recentCount: number;
}